// matico_charts/src/_archive/components/types.ts
export type MarginSpec = {
  top: number;
  right: number;
  bottom: number;
  left: number;
}

export type Domain2D = {
  x0: number;
  x1: number;
  y0: number;
  y1: number;
}
// data
export type DataRow = { [property: string]: any };
export type DataCollection = Array<DataRow>;
// ...
export interface StaticMapSpec extends BaseLayerSpec {
  proj: string;
  rotation?: number;
  fill?: ColorFunction | ColorOutput;
  //background?: ColorOutput;
  gratOn?: boolean;
  gratColor?: ColorOutput;
  strokeWidth?: number;
  strokeColor?: ColorOutput;
  pointRadius?: number;
  events?: boolean;
}

export interface ScatterSpec extends BaseLayerSpec {
  color?: ColorFunction | ColorOutput;
  scale?: number | SizeFunction;
  shape?: string | ShapeFunction;
  onHover?: InteractiveFunction;
  onSelect?: InteractiveFunction;
  tooltip?: boolean; // TooltipSpec;
}

export interface LineSpec extends BaseLayerSpec {
  lineFunction?: LineFunction;
  lineColor?: string;
  lineWidth?: number;
}

export interface BarSpec extends BaseLayerSpec {
  padding?: number;
}

export interface PieSpec extends BaseLayerSpec {
  reverseSort?: boolean;
  innerRadius?: number;
  padding?: number;
}

export interface HeatmapSpec extends BaseLayerSpec {
  xBins: number,
  yBins: number,
  binnedData?: DataCollection[]
}

export type LayerSpec = ScatterSpec | LineSpec | BarSpec | PieSpec | HeatmapSpec | StaticMapSpec;

// layouts
export type OverwriteProperty = { [property: string]: any };
export interface AxisSpec {
  position: YAxisPosition | XAxisPosition;
  scaleType: ScaleType;
  display?: boolean;
  scaleParams?: OverwriteProperty;
  tickFormatFunc?:  (d: number) => string;
}

export interface GridSpec {
  columns: boolean;
  rows: boolean;
}
export type BooleanOrAxisSpec = boolean | AxisSpec;
export interface WidthAndHeight {
  width: number;
  height: number;
}

//interactions
export interface BrushSpec {
  horizontal: boolean;
  vertical: boolean;
}
export type BooleanOrBrushSpec = boolean | BrushSpec;
export interface Domain {
  x0: number;
  x1: number;
  y0: number;
  y1: number;
}
export type xyDomainCallback = (d: Domain) => void;

export interface ChartSpaceSpec {
  data: DataCollection;
  layers?: Array<LayerSpec>;

  title?: string;
  subtitle?: string;
  attribution?: string;

  dimensions?: WidthAndHeight;
  useBrush?: BooleanOrBrushSpec;
  onBrush?: xyDomainCallback;

  margins?: Partial<MarginSpec>;

  updateTrigger?: any;
}